switch to int

+++ Models/Playlist.cs
@@ -2,7 +2,7 @@ namespace tourmaline.Models;
 
 public class Playlist
 {
-    public Playlist(long id = -1, string username = "", string name = "", string description = "")
+    public Playlist(int id = -1, string username = "", string name = "", string description = "")
     {
         Id = id;
         Name = name;
@@ -11,7 +11,7 @@ public class Playlist
         Songs = new List<Song>();
     }
 
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Username { get; set; }
     public string Name { get; set; }
     public List<Song> Songs { get; set; }
