+++ Models/Song.cs
@@ -2,7 +2,7 @@ namespace tourmaline.Models;
 
 public class Song
 {
-    public Song(long id = -1, string uploader = "", string name = "",
+    public Song(int id = -1, string uploader = "", string name = "",
         string description = "", List<string>? tags = null, int listenTimes = 0, int favorites = 0)
     {
         Id = id;
@@ -15,7 +15,7 @@ public class Song
         Favorites = favorites;
     }
 
-    public long Id { get; set; }
+    public int Id { get; set; }
     public DateTime UploadTime { get; set; }
     public string Uploader { get; set; }
     public string Name { get; set; }
